requirements.txt
Django==5.1.6
djangorestframework==3.15.2
# to start the django server with ssl
django-extensions==3.2.3
Werkzeug==2.0.3
# http library for Python
requests==2.26.0
# handles env for python
environs==9.3.2
# handles env for django
django-environ==0.4.5
# postgres adaptator for Python
#psycopg2==2.9
psycopg2-binary==2.9
# Hashing Passwords with Argon2
argon2-cffi==23.1.0
# packages for JWT tokens
djangorestframework-simplejwt==5.5.0
django-cors-headers==3.7.0
# server for asynchronous requests and handling websockets
daphne==3.0.2
# asynchronous Django tool to handle websockets and real-time features
channels==3.0.5
# channels[daphne]
# helping with the coordination of websockets for Django Channels
channels-redis==3.2.0
# necessary for wss
cryptography==39.0.0
# whitenoise==6.9.0
Pillow==11.1.0
# package for OTP(2FA)
pyotp==2.6.0
# Handles 0Auth2 authentication
django-oauth-toolkit==3.0.1
requests-oauthlib==2.0.0
# for online status
redis==5.2.1
# better compatibility with asynchronous server like daphne
aioredis==1.3.1
# for api-cli
websockets==10.4
keyboard==0.13.5
readchar==4.2.1

